feat(theme-builder): move utils to a dedicated subdirectory
