+++ packages/@poupe-theme-builder/src/core/__tests__/colors.test.ts
@@ -1,6 +1,6 @@
 import { expect, test, it, describe } from 'vitest';
 import { Hct } from '@material/material-color-utilities';
-import { uint32 } from '../utils';
+import { uint32 } from '../../utils/utils';
 
 import {
   type Color,
