--- packages/@poupe-theme-builder/src/server.ts
@@ -1,1 +1,0 @@
-export * from './server/index';
