--- packages/@poupe-theme-builder/src/utils.ts
@@ -1,1 +1,0 @@
-export * from './core/utils';
